requirements.txt
aiohttp==3.8.4
aiolimiter==1.0.0
aiosignal==1.3.1
anyio==3.6.2
APScheduler==3.10.0
async-timeout==4.0.2
attrs==22.2.0
CacheControl==0.12.11
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
cryptography==39.0.1
frozenlist==1.3.3
google-api-core==2.11.0
google-api-python-client==2.79.0
google-auth==2.16.1
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.10.0
google-cloud-storage==2.7.0
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.58.0
grpcio==1.51.3
grpcio-status==1.51.3
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.16.3
httplib2==0.21.0
httpx==0.23.3
hyperframe==6.0.1
idna==3.4
jmespath==1.0.1
msgpack==1.0.4
multidict==6.0.4
pendulum==2.1.2
proto-plus==1.22.2
protobuf==3.20.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydnevnikruapi==0.1.1
PyJWT==2.6.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.1
python-telegram-bot==20.1
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
pytzdata==2020.1
requests==2.28.2
rfc3986==1.5.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
socksio==1.0.0
tornado==6.2
tzdata==2022.7
tzlocal==4.2
uritemplate==4.1.1
urllib3==1.26.14
yarl==1.8.2
